+++ source/Sashimi.Azure.Accounts.Tests/Sashimi.Azure.Accounts.Tests.csproj
@@ -9,6 +9,7 @@
     <PackageReference Include="nunit" Version="3.12.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="3.15.1" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.4.0" />
+    <PackageReference Include="Octopus.Server.Tests.Extensibility" Version="11.0.0" />
     <PackageReference Include="TeamCity.VSTest.TestAdapter" Version="1.0.22" />
   </ItemGroup>
