+++ source/Sashimi.Azure.Accounts/AccountTypes.cs
@@ -5,6 +5,5 @@ namespace Sashimi.Azure.Accounts
     public static class AccountTypes
     {
         public static readonly AccountType AzureServicePrincipalAccountType = new AccountType("AzureServicePrincipal");
-        public static readonly AccountType AzureSubscriptionAccountType = new AccountType("AzureSubscription");
     }
 }
