+++ source/Sashimi.Azure.Accounts/AzureServicePrincipalAccountDetails.cs
@@ -8,7 +8,7 @@ using Sashimi.Server.Contracts.Variables;
 
 namespace Sashimi.Azure.Accounts
 {
-    public class AzureServicePrincipalAccountDetails : AccountDetails, IExpandVariableForAccountDetails
+    class AzureServicePrincipalAccountDetails : AccountDetails, IExpandVariableForAccountDetails
     {
         public override AccountType AccountType { get; } = AccountTypes.AzureServicePrincipalAccountType;
